requirements.txt
appdirs==1.4.0
Cerberus==0.9.2
click==6.6
conda==4.1.11
conda-env==2.5.0a0
decorator==4.0.10
Eve==0.6.5.dev0
Events==0.2.1
fcl==0.1
Flask==0.10.1
Flask-PyMongo==0.4.1
ipykernel==4.3.1
ipython==5.0.0
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.3.0
jupyter-console==5.0.0
jupyter-core==4.1.0
MarkupSafe==0.23
MeshPy==2016.1
mistune==0.7.2
mpmath==0.19
msgpack-python==0.4.8
nbconvert==4.2.0
nbformat==4.0.1
networkx==1.11
notebook==4.2.2
numpy==1.11.1
ordereddict==1.1
path.py==0.0.0
pexpect==4.0.1
pickleshare==0.7.3
prompt-toolkit==1.0.3
ptyprocess==0.5.1
py==1.4.31
pycosat==0.6.1
pycrypto==2.6.1
pyglet==1.2.4
Pygments==2.1.3
pymongo==3.3.0
pytest==3.0.0
python-dotenv==0.5.0
pytools==2016.2.1
PyYAML==3.11
pyzmq==15.3.0
qtconsole==4.2.1
requests==2.10.0
Rtree==0.8.2
ruamel-yaml===-VERSION
scipy==0.18.0
Shapely==1.5.16
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
svg.path==2.1.1
sympy==1.0
terminado==0.6
tornado==4.4.1
traitlets==4.2.2
trimesh==1.15.16
wcwidth==0.1.7
Werkzeug==0.11.3
